Fixed namespace after fodler renaming.
